moved function into service
